Fix version bump type detection [no version bump]
The version bump type can be controlled by keywords in [] in commit messages.
This wasn't working properly because `master` was being checked out too early
to properly detect the commit history for the changes. This moves the `master`
checkout to after the version bump type detection code to avoid the issue.